// src/app/components/directors-office/director-components/add-item-modal/model-interface.ts
export interface modelADD {
  name: string;
  SToplata: string;
  STzaiavki: string;
  employee: string;
  vipiska: string;
  company: string;
  adres: string;
  opisanie: string;
  nomer: string;
  sale_zatrat: number;
  obfaia_sale: number;
}

export interface StatusApplication {
  status_application: string;
}
export interface StatusPayment {
  status_payment: string;
}